+++ directProblemQuadratureMethod/Sources.h
@@ -11,8 +11,8 @@ struct Source
 	const size_t numberSource = 5;
 	// координаты источников на двухмерной плоскости
 	const std::vector<Point> node = {
-		{ -0.1, 0.0 }, { -0.1, 0.2 }, { -0.1, 0.4 }, { -0.1, 0.6 }, { -0.1, 0.8 }, { -0.1, 1.0 } };
-	std::complex<double> Function(const Point source, const double x, const double y) const;
+		{ -0.1f, 0.0f }, { -0.1f, 0.2f }, { -0.1f, 0.4f }, { -0.1f, 0.6f }, { -0.1f, 0.8f }, { -0.1f, 1.0f } };
+	std::complex<float> Function(const Point source, const float x, const float y) const;
 };
 
 // печать значений источника в файл "Source.txt"
