+++ directProblemQuadratureMethod/directProblem_utils.h
@@ -2,18 +2,18 @@
 
 #include "Sources.h"
 
-void GetSubstantiveMatrix(const std::vector<std::vector<std::vector<std::vector<std::complex<double>>>>> & a,
-	const std::vector<std::vector<std::complex<double>>> & b,
-	const std::vector<std::vector<double>> & xi,
-	std::vector<std::vector<std::complex<double>>> & substantiveMatrix);
+void GetSubstantiveMatrix(const std::vector<std::vector<std::vector<std::vector<std::complex<float>>>>> & a,
+	const std::vector<std::vector<std::complex<float>>> & b,
+	const std::vector<std::vector<float>> & xi,
+	std::vector<std::vector<std::complex<float>>> & substantiveMatrix);
 
 void GetRightPartEquation(const Source & source, size_t count,
-	std::vector<std::complex<double>> & rightPartEquation);
+	std::vector<std::complex<float>> & rightPartEquation);
 
-void InverseRenumbering(const std::vector<std::complex<double>> & numbered_u,
-	std::vector<std::vector<std::complex<double>>> & u);
+void InverseRenumbering(const std::vector<std::complex<float>> & numbered_u,
+	std::vector<std::vector<std::complex<float>>> & u);
 
 void GetOverlineU(const Source & source, size_t count, 
-	const std::vector<std::vector<std::vector<std::complex<double>>>> & overline_a,
-	const std::vector<std::vector<double>> & xi, const std::vector<std::vector<std::complex<double>>> & u,
-	std::vector<std::complex<double>> & overline_u);
+	const std::vector<std::vector<std::vector<std::complex<float>>>> & overline_a,
+	const std::vector<std::vector<float>> & xi, const std::vector<std::vector<std::complex<float>>> & u,
+	std::vector<std::complex<float>> & overline_u);
