+++ directProblemQuadratureMethod/exact_solution.cpp
@@ -4,21 +4,21 @@
 
 using namespace std;
 
-void GetExactSolution(vector<vector<double>> & xi)
+void GetExactSolution(vector<vector<float>> & xi)
 {
-	double sigma = 25;
+	float sigma = 25.0f;
 	for (size_t i = 0; i <= NUMBER_PARTITION_POINT; ++i)
 	{
 		for (size_t j = 0; j <= NUMBER_PARTITION_POINT; ++j)
 		{
-			xi[i][j] = exp(-((i * h - 0.6) * (i * h - 0.6) + (j * h - 0.6) * (j * h - 0.6)) * sigma);
+			xi[i][j] = exp(-((i * h - 0.6f) * (i * h - 0.6f) + (j * h - 0.6f) * (j * h - 0.6f)) * sigma);
 			//xi[i][j] = 0.8 * exp(-((i * h - 0.8) * (i * h - 0.8) + (j * h - 0.8) * (j * h - 0.8)) * sigma) +
 				//0.2 * exp(-((i * h - 0.2) * (i * h - 0.2) + (j * h - 0.2) * (j * h - 0.2)) * sigma);
 		}
 	}
 }
 
-void WriteSolutionFile(vector<vector<double>> & xi)
+void WriteSolutionFile(vector<vector<float>> & xi)
 {
 	ofstream file_xi("exact_xi.txt");
 	file_xi << fixed << setprecision(6);
