+++ directProblemQuadratureMethod/exact_solution.h
@@ -1,5 +1,5 @@
 #pragma once
 
-void GetExactSolution(std::vector<std::vector<double>> & xi);
+void GetExactSolution(std::vector<std::vector<float>> & xi);
 
-void WriteSolutionFile(std::vector<std::vector<double>> & xi);
+void WriteSolutionFile(std::vector<std::vector<float>> & xi);
